+++ connectors/tests/test_sync.py
@@ -14,7 +14,7 @@ from aioresponses import CallbackResult
 
 from connectors.byoc import DataSourceError, SyncJobIndex
 from connectors.conftest import assert_re
-from connectors.services.sync import SyncService
+from connectors.services.scheduling import SchedulingService
 from connectors.tests.fake_sources import FakeSourceTS
 
 CONFIG = os.path.join(os.path.dirname(__file__), "config.yml")
@@ -187,7 +187,7 @@ class Args:
 
 
 def create_service(config_file, **options):
-    return SyncService(Args(config_file, **options))
+    return SchedulingService(Args(config_file, **options))
 
 
 def test_bad_config():
